refactor(doctor): split Codex route diagnostics
