+++ scripts/lib/session-accessor-debt-baseline.json
@@ -8,7 +8,7 @@
     "src/auto-reply/reply/commands-session-store.ts": 2,
     "src/commands/doctor-heartbeat-session-target.ts": 2,
     "src/commands/doctor-state-integrity.ts": 2,
-    "src/commands/doctor/shared/codex-route-warnings.ts": 3,
+    "src/commands/doctor/shared/codex-route-session-repair.ts": 3,
     "src/config/sessions/session-accessor.ts": 12,
     "src/config/sessions/store-load.ts": 3,
     "src/config/sessions/store.ts": 10,
@@ -21,7 +21,7 @@
     "src/commands/doctor-session-snapshots.ts": 2,
     "src/commands/doctor-session-state-providers.ts": 2,
     "src/commands/doctor-state-integrity.ts": 2,
-    "src/commands/doctor/shared/codex-route-warnings.ts": 2,
+    "src/commands/doctor/shared/codex-route-session-repair.ts": 2,
     "src/gateway/test-helpers.mocks.ts": 1,
     "src/infra/state-migrations.session-store.ts": 2,
     "src/plugins/registry-runtime.ts": 1
